Dashboard_App/interactive-backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
guardrails-ai==0.6.5
guardrails-api-client==0.4.0a1
guardrails_hub_types==0.0.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openai==1.72.0
pandas==2.2.3
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
python-dotenv==1.1.0
Werkzeug==3.1.3
